ASACORE-775 About Feature rework - separate the icon
This separates the About Icon into its own class.  This solves several
problems.  This solves the problem that we can not return errors when
the content passed into the AboutIconObj could not be marshaled.

Rather than exposing each of the org.alljoyn.Icon methods the user
calls GetIcon and this will fill in all of the values that are found
in the Icon.

Updated the test programs and unit tests.

Also as part of this change the AboutService and AboutClient found in
the test folder were renamed to aservice aclient to help distinguish
the test programs from the sample programs.

Change-Id: I3edf9bdc60ef01513e3e1f540a92afc72b78cdff
Signed-off-by: George Nash <georgen@qce.qualcomm.com>