ASACORE-47 Improve Properties Change notification filtering
If 2 remote devices implement the same interface on the same object
path, then the ProxyBusObject instanciated for only one of those
devices' objects would get Properties Changed events from both remote
devices and the application would not be able to determine whether the
received Properties Changed event was related to the specific instance
of ProxyBusObject or not.

This change adds code to collect the unique bus name containing the
BusObject the ProxyBusObject is proxying so that it can filter out
Properties Changed events from other remote devices.

Change-Id: I6e393ad204b84f2626359693bd4876269691a9ee
Signed-off-by: Steve Kinneberg <stevek@qce.qualcomm.com>