ASACORE-996 Make test programs compatible with 14.06 code
the aservice and aclient program are used to verify backwards compatibility
with the existing about service.  The existing about service expects certain
interfaces to be advertised so these programs need to be updated to work
with those advertisements.

Updated the printf statements to print a user friendly version of the AboutData
and ObjectDescription.

Change-Id: I14727924ba902ec06800021bfedd549f82427e59
Signed-off-by: George Nash <georgen@qce.qualcomm.com>