ASACORE-775 About feature rework
Fixed the Constructor that Fills in the fields of the AboutData
from the MsgArg.  The user could pass in a MsgArg that is not the
default language for that reason the constructor requires a language
parameter.

Fixed issue that the m_supportedLanguages member variable was not
properly initialized when creating the the AboutData class from a
MsgArg.

Add the GetFields member function to the AboutData class. This is
used to get a list of all the fields that came in the Announce signal
this is the only way to get a list of all of the fields that are found
in the AboutData class.

Change-Id: I12e5da50f445803043faa409c5a78d1cf4cded3c
Signed-off-by: George Nash <georgen@qce.qualcomm.com>