ASACORE-775 About Feature: Proxy objects Inherit ProxyBusObject class
There is not need to make the ProxyObject a member of the About
ProxyObjects.  All they needed to do was inherit the ProxyBusObject
class.  This did not change the way the code it used it just
removed the need to have an ProxyBusObject as a class member.

Removed the MessageReceiver class from the list of inherited classes
in the AboutProxy class. It was not used in the AboutProxy class and
was not needed.

Also fixed a build warning in the AboutClient sample.

Change-Id: Iee227449dafc655da939e4c34e5880d546271a4e
Signed-off-by: George Nash <georgen@qce.qualcomm.com>