ASACORE-1007 About Feature: Fixed WhoImplements with NULL value
The method WhoImplements improperly handled NULL value. It returned
ER_BAD_ARG_1 when it should passed it on to the lower member function
that was already designed to handle NULL value properly.

Renamed the AboutDataTest group name from AboutData to AboutDataTest
this fits in with the other group names in the unit tests.  They all
end in the word Test.

Change-Id: Ie931af420264aa2d04e5d60723d3de48f0af6b74
Signed-off-by: George Nash <georgen@qce.qualcomm.com>