ASACORE-775 About Feature: error checking code improvements
The error checking code in the AboutObj.Announce function were
modified so it does not make as many calls into the
AboutDataListener.  the error checking code would call the
functions in the AboutDataListener a total of 5 times.

If the AboutDataListener was getting the AboutData from a
file this could be a really slow operation. With these
modifications the abautDataListener is only called twice
once to get the AboutData and again to get the AnnouncedData

Change-Id: I5819f450003265cd6f72bc0f3dc83ebf4540369d
Signed-off-by: George Nash <georgen@qce.qualcomm.com>