ASACORE-775 Add the function BusObject::SetAnnounceFlag
The fact that the AddInterface method is a private member function
has caused difficulty for the Objective-c language bindings that
use a code generator to create a BusObject for users.  The fact
that the AddInterfaces means users must modify the generated code
to announce an interface. The only other time the user must modify
the generated code is when working with variant MsgArgs so this is
unusual. There has also been some argument that this would be useful
to have in the Java language bindings.

Since the Announce flag is not required till the announce method is
called it can be changed using the SetAnnounceFlag. Any time the
announce flag is changed AboutObj::Announce should be called to
inform the bus of the change.

Change-Id: I70c9a9aa57ed7a22827e65e5f85c6c7ea668547e
Signed-off-by: George Nash <georgen@qce.qualcomm.com>