ASACORE-1079: Enable Windows app-to-app isolation
The check-in implements application isolation for universal applications on
Windows. Universal applications are not allowed to communicate with other
univeral apps on the same machine, nor can they talk with desktop applications
on the machine. This framework allows us to define policy rules on Windows
to prevent AllJoyn from violating these policies.

Change-Id: I6ac7e63ce421e45d34f5a42a60921f9043b548fd
Signed-off-by: Rob Smith <robs@microsoft.com>