ASACORE-1014 : Named pipes support for Windows
This feature is to implement named pipe transport between leaf node
and router node of AllJoyn when both of them are on the same Windows
system. This feature is only for the Windows 10 and will not work on
previous Windows releases.

Iteration 2: Updating the copyright information for newly added files.

Change-Id: If3567101bbc09e5b420412fa82b135b5dbe5d948
Signed-off-by: Rajneesh <rajban@microsoft.com>