ASACORE-1056 Fixed memory leak
ASACORE-1056 Fixed memory leak found in MethodTable::Add

The Add function was adding entries to the hashMap based on
a Key generated from the object path, interface name, and method
name.

The Add function  also adds an additional entry that intentionally
ignored the interface name.  If you have multiple interfaces that
have the same object path and method name then only one entry would
be written into the hashMap instead of multiple entries for each
interface.  The entry must make it into the hashMap or the code
responsible for freeing the memory can not find the entry to free
the entries memory.

The specification states if there are multiple properties on an
object with the same name, the results are undefined. So if for
this implementation the name is only added to the hashMap if
the name is not already found. This prevents the memory leak
since it prevents creation of the new entry.

Change-Id: Ie5125a2f3d968785ca1dae86e0a44dad7e97fb5c
Signed-off-by: George Nash <georgen@qce.qualcomm.com>