ASACORE-2642 Don't install all test programs in the SDK
Most of the functionality the test programs have been duplicated
in the unit tests or are not used to regularly test and verify
AllJoyn functionality.  These programs are installed in the test
directory not in the dist directory since they don't need to be
part of the SDK.

The list of the programs are roughly in alphabetic order. This
made comparing the list programs easier while updating the
SCons file.

Corresponding changes made in reference SDK TOC files.

Change-Id: Ic5f4c4899d4b4424d25fefd76996845fb03cfe94
Signed-off-by: George Nash <georgen@qce.qualcomm.com>
Signed-off-by: Tony Wallace <awallace@qce.qualcomm.com>