ASACORE-1130 AboutData AppId must be 128-bits
When setting the AppId check that the number of bytes is 16
making sure the AppId is 128-bits.

this adds the status:
ER_ABOUT_INVALID_ABOUTDATA_FIELD_VALUE

I also changed the status
ER_ABOUT_INVALID_ABOUT_DATA_LISTENER to
ER_ABOUT_INVALID_ABOUTDATA_LISTENER

This was done to match the spacing used in other status
values.

Change-Id: I0f839c6df092ce9f5a126f5245c897afa2d4a7fc
Signed-off-by: George Nash <georgen@qce.qualcomm.com>