ASACORE-1130 AboutData AppId must be 128-bits
Since the AppId must be a UUID as defined by RFC 4122
Its resonable to expect a user to want to set the AppId
using a string as specified in RFC 4122.  Since I already
had some code that parsed a HEX string without the '-' character
I just made a method that would parse the with the '-' character
this make since to have since the original About Service code
let the develper set the AppId using a string.  This Also
improves parsing making sure the value entered in using the
CreateXml function is also checked.

Change-Id: I18ae551966ef150ef832e326b34107a75f5029e5
Signed-off-by: George Nash <georgen@qce.qualcomm.com>