ASACORE-1140 Turn off advertisement timer when V1 traffic is disabled
The periodic timer for retransmitting the queued advertisements can
be turned off when V1 traffic is not enabled (which is the default).
The V1 traffic setting is changeable through the config flag:
ns_enable_v1.

Without this change, IpNameServiceImpl::DoPeriodicMaintenance gets
called every 1 second when there's a queued advertisement but nothing
is sent out, impacting device power unnecessarily.

Change-Id: I0df5ccab11be9cee2ad935f4124cd3f0d04a80d8
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>