ASACORE-1033 Modified About Proxy unit tests
Re-enabled mulitiple AboutListener test because we are no longer
able to reproduce the issue.

Also:

Modified the About Proxy unit tests to not use wait/notify
to handle sycronisation since there appears to be some question if
it is properly causing the code to wait for the event.  I replaced
the code with a busy wait loop similar to the wait loops found in
the C++ unit tests. These are useful because they don't need
to be syncronised between threads.  We just check if an event has
happend if it has not we sleep for a short while and check again
until a max time out.

This change was done while investigating  ASACORE-1138 issue
this is not a fix for that issue  but should reduce syncronisation
errors.

Change-Id: Ib7421ded5a675c4135be86e2526826d177891764
Signed-off-by: George Nash <georgen@qce.qualcomm.com>