ASACORE-1138 Fixed issue causing unit test failures
The Java AboutObj was being garbage collected when it was still
needed by the C++ code.  When an announcement is sent out
we grab a GlobalRef to the Java AboutObj if it has not already
been obtained by a previous announce call.

The GlobalRef is held till eather the BusAttachment is disconnected
or till cancelAnnouncement is called.

Change-Id: I3cc508424a809b5dcccb7e0a6c8c44803c12c2d8
Signed-off-by: George Nash <georgen@qce.qualcomm.com>