ASACORE-1211 Language tags should be case insensitive
This adds a case insinsitive comparator to the AboutData class
so that when fetching from the localizedPropertyStore map the
case of the language tag is unimportant. Case insinsitive
comparator added to the language tags set so the case of the
language tag is unimportant when adding the language.

Also added case insensitivity to the java AboutDataListeners

Change-Id: If63fe4289e2ea4b6a7ea87a82c558ea4e9050b25
Signed-off-by: George Nash <georgen@qce.qualcomm.com>