ASACORE-1229 AppId can be larger or smaller than 128-bits
Let the user set the AppId to a value that is not 128-bits

Untill recently the requirement to have the AppId be 128-bits
was not clearly stated. This lead to developers creating
applications that had AppIds larger or smaller than 128-bits.

So that we don't break applications that are using a non-128-bit
AppId we will still send out an annouce signal that is not
128-bits.

If a user set the AppId to a value that is not 128-bits the
AboutData::SetAppId and the AboutObj::Announce member functions
will return ER_ABOUT_INVALID_ABOUTDATA_FIELD_APPID_SIZE status
if the AppId is not 128-bits.  The developer can choose to ignore
this warning.

Change-Id: I099333c8e62edae5c5c78efc7e1d7dcc4e720996
Signed-off-by: George Nash <georgen@qce.qualcomm.com>