ASACORE-1119 Unit test should be using GUID::ToString function
The unit test was using the GUID::RenderByteString member function
which literally changes a byte array to a string array no conversion.

this meant if the first byte was `0` then the RenderByteString looked
like a NULL and would result in the code segfaulting.  The test code
should have use the GUID::ToString member function.  This converts
the array of bytes to a a string representation of the bytes. so a
'0' byte becomes "00" string.

Change-Id: I541cfff6353ebf3611d766e8ee4ea8c791fd33aa
Signed-off-by: George Nash <georgen@qce.qualcomm.com>