ASACORE-1279: Fix AddMatch rules for PropChange when bus restarts
Fix a problem where internal accounting of AddMatch rules for
PropertiesChagned gets out of sync with the router if the bus is
disconnected then reconnected but the ProxyBusObject stays live across
the disconnect and reconnect.

ASACORE-1304: Fix Java method return type

Return type for unregisterPropertiesChangedListener in
ProxyBusObject.java should be Status and not void.

Change-Id: Ib6a0e22d592c046c331637092dddfcfceeb10d53
Signed-off-by: Steve Kinneberg <stevek@qce.qualcomm.com>