ASACORE-1322: StaticGlobalsInit is defined is a public C header
Add an #ifdef around the StaticGlobalsInit and define it by default in the
AllJoyn core.

We just discovered a blocking issue bringing this change internally. The
StaticGlobalsInit class is in the platform.h header which is used by both
the C and C++ APIs, so when we try to build our apps on the C API they try
to build in the StaticGlobalsInit. The StaticGlobalsInit class shouldn't be
part of the public API and isn't needed by these apps, so they end up
hitting a linker error.

The correct solution is to define this StaticGlobalsInit class in a separate
header that isn't included in the C API, but since 14.12 is about to be
released we came up with a quicker fix: wrap the class definition in #ifdefs
and exclude it by default. In the AllJoyn SConscript file we can enable
building it, but this way when others try to build apps from the public
headers the class is skipped.

Change-Id: Ia981ae395e1cb03f2de9e102501e8dfae5a36bc7
Signed-off-by: Rob Smith (WINDOWS) <robs@microsoft.com>