ASACORE-836 Fixed stack overflow error in MsgArg tests
When the large arrays were inside the TEST the test framework
tries to pass the array on the stack.  This resulted in a
stack overflow error on some systems.  The way to solve the
issue was to pull the array out of the TEST so it was not
sent on the stack.

Change-Id: Ie9e0e343931feb821363c61c0ad11f835e888fec
Signed-off-by: George Nash <georgen@qce.qualcomm.com>