ASACORE-1382: Windows app isolation policies prevent an application from communicating with itself
The Windows universal app isolation policies require that universal apps cannot talk to each other,
however this rule was also preventing apps from communicating with themselves (for apps with
multiple busattachments).

The fix requires a new type of rule to be added to the policydb: a special = rule type which will
kick in if the user or group ID of the sender and receiver are equal. I've also added code to
populate the userId for universal apps to a unique ID per-process.

Here are the rules we're using so you can better understand:

 <policy group="UniversalWindowsApplication">
 <deny send_group="UniversalWindowsApplication"/>
 <deny receive_group="UniversalWindowsApplication"/>
 <deny send_group="DesktopApplication"/>
 <deny receive_group="DesktopApplication"/>
 <allow send_usersequal/>
 <allow receive_usersequal/>
 </policy>
 <policy group=\"DesktopApplication\">
 <deny send_group="UniversalWindowsApplication"/>
 <deny receive_group="UniversalWindowsApplication"/>
 </policy>

Change-Id: I587d31d602e9c8b7a18ce4660f45e4ece309d58a
Signed-off-by: Rob Smith (WINDOWS) <robs@microsoft.com>