ASACORE-1026 Update uncrustify to the latest version v0.61
The Goal was to make sure builds did not break if they have not
updated to the latest version of uncrustify. Since the newer version of
uncrustify have fixed some issues like properly detecting pointers and the ':'
character in case statements some white space differences were unavoidable.

To handle the differences two uncrustify config files were created.  When possible
the white space policy that was buggy was turned off in the old config file while
keeping it on in the newer version of the config file.

In situations that there were disagreements that could not be ignored by changing
an option in the config file that file was added to the ignore list.  The larger
the file the more likely it is to have an unavoidable issue. like alljoyn_java.cc
and BusAttachment.cc.

This approach has some advantages and disadvantages.
Advantages:
  - We have a way to update the whitespace tools
  - The code is not being change even if you have different versions of uncrustify
  - We have a way to address whitespace related bugs found in older versions of
    uncrustify
  - We can start using features found in newer versions of uncrustify

Disadvangages:
  - Must have a different config file for each version of uncrustify supported
  - Older versions of uncrustify may not find all the whitespace errors
  - There will still be a minor amount of code churn each time a new version
    of uncrustify is released.

The code changes now work with both uncrustify v0.57 and uncrustify v0.61 with
a little work other versions of uncrustify could be added in.

Change-Id: Ibce782865572c8a77008e3d561fcd470ee1aec61
Signed-off-by: George Nash <georgen@qce.qualcomm.com>