ASACORE-1532 ASACORE-1789: Rework DaemonRouter::PushMessage()
Simplified the implementation of DaemonRouter::PushMessage() to
improve understandability and maintainability.  This includes some
isolated compatibility hacks so that other code that uses or is used
by it will still function normally.  These hacks can be easily removed
once the correlated code that needs those hacks are fixed to work with
the simplified behavior.

Some minor additions have been made to RuleTable and NameTable to
reduce the connectivity between DaemonRouter and those classes.
DaemonRouter::PushMessage() no longer has any knowledge about the
inner workings of those classes.

1798: Also adds a call to DaemonRouter::RemoveSessionRoutes(SessionId) from
AllJoynObj that is required for the DaemonRouter::PushMessage rework.

Change-Id: Ie17096514a70b553eb983520e09c460a18ed3514
Signed-off-by: Steve Kinneberg <stevek@qce.qualcomm.com>
Signed-off-by: Sheshambika Venkateshwaran <sheshamb@qce.qualcomm.com>