ASACORE-1417: Send SLS request signal to unchanging name.
It's possible for the SLS producer to change their well-known name
between the consumer joining and sending the request signal.  In that
case, the request signal is dropped since the destination no longer
exists.  The solution is to use the unique name that is the owner of the
well-known name instead of the well-known name.

Change-Id: I94c4466c869a893f74986bb0c2f7dc24130853f7
Signed-off-by: Todd Malsbary <tmalsbar@qce.qualcomm.com>