ASACORE-1494: Use Windows crypto APIs to protect AllJoyn key store
Use the Windows CryptProtectData APIs to encrypt the key store. As part of
this change I move the keystore to a folder with a different name of Windows
as this change will not migrate old key stores to new encrypted key stores.
Any application currently using security will require users to re-authenticate
when they upgrade the application from 14.12 to 15.04. Currently there are no
known AllJoyn applications on Windows using authentication so we believe the
user impact will be minimal.

This change also added a new function to the FileSource to get the file size.

Change-Id: I0cf8292f32094ef7254a1b5ee911c19115d6365f
Signed-off-by: Rob Smith (WINDOWS) <robs@microsoft.com>

Fix code review error in KeyStoreListener

Change-Id: Ic5c87ec851a80fddacb82b1434a53dfdabceeaa2
Signed-off-by: Rob Smith (WINDOWS) <robs@microsoft.com>