AutoPinger fixes (mostly locking related)
The AutoPinger could not guarantee that the lock ordering between the
global ctxLock and the per-pinger locks was always respected. The
simplest solution was to revert to a global lock for all AutoPinger
instances. Given the relatively low frequency of AutoPinger-related
actions, we believe the global lock will not impair performance.

PingListener callbacks are now also invoked without any locks held, to
avoid possible lock ordering problems between application locks and the
global pinger lock. As a result, ~AutoPinger and
AutoPinger::RemovePingGroup now busy-wait for all callbacks to be
terminated before doing their job.

The final change is a responsiveness improvement. When a new destination
was added to a ping group, the AutoPinger waited until the next ping
interval before doing its initial ping call to that destination.
That means applications potentially had to wait for confirmation of the
existence/nonexistence of a destination for the entire ping interval.
For long ping intervals (which really ought to be the case if you're
using AutoPinger properly), that is too long. Therefore, the AutoPinger
now performs an out-of-sequence ping for a new destination directly when
it is added to a ping group.

Change-Id: I6ad8bcc5e372ebdda27781f6b39e0d8a60583a58
Signed-off-by: Meggan Ehret <meggan@qeo.org>