ASACORE-1516 Add a PermissionMgmtListener
In addition to adding a PermissionMgmtListener member functions have
been added to BusAttachment to register and unregister the listener.
Member functions added to make calling the Add match rule for the
NotifyConfig signal simpler.

The PermissionMgmtListener currently returns MsgArg for the publicKeyInfo,
trustAnchors, and memberships. These should be converted from a MsgArg to
a C++ type in a future commit. This is to get the listener in place so it
can be expanded in a later commit.

Change-Id: I036e0b27d3d2acae55fa6d71a149796e97723ac3
Signed-off-by: George Nash <georgen@qce.qualcomm.com>