ASACORE-1518 add PermissionMgmtProxy class
This class replaces many of the helper functions found in the unit
test PermissionMgmtTestHelper.

While doing the work for this change I changed the ProxyBusObject's
internal BusAttachment pointer to a protected member variable. While
comparing the work being done for the PermissionMgmtProxy and the
the work done for the AboutProxy I realized they both were keeping
a reference to the BusAttachment the ProxyObject was associated with
so that it could be used when creating the reply Message.  This is
a pattern that seems so common that it will be used over and over.
Since the ProxyBusObject already has a BusAttachment reference why
add an additional reference for every ProxyBusObject that extends the
ProxyBusObject.

Change-Id: I8bb771a108b4144a36a2a8d64ac7e3bf8f35df39
Signed-off-by: George Nash <georgen@qce.qualcomm.com>