ASACORE-1551 Use BCRYPT_SUCCESS macro to check return value of BCrypt functions
The previous change under this bug left us with a mixture of use of BCRYPT_SUCCESS and explicit comparisons of NTSTATUS values to zero in the CNG wrapper code.

Just for consistency's sake, this change goes through all the rest of the latter and changes them to use the BCRYPT_SUCCESS macro. Any if statement that checked for an NTSTATUS < 0 is now !BCRYPT_SUCCESS, and NTSTATUS >= 0 is now BCRYPT_SUCCESS.

Change-Id: If9ad6bff833599da1584bce93288c9a2d74af08a
Signed-off-by: Kevin Kane <kkane@microsoft.com>