ASACORE-1595 Set default RSA keylength to 2048
In the RSA auth mech, if a certificate is not provided by the
application or found in the keystore, the library will generate a
self-signed certificate. When a self-signed certificate is generated,
the library hardcodes a 512-bit keylength. Given a 512-bit RSA public
key it is possible to recover the secret key with a small amount of
computing resources in a reasonable amount of time. This allows a
passive network adversary to recover the session key, then
subsequently use the private key to authenticate as the peer (who
originally generated the 512-bit key). This change replaces "512"
with a constant, and sets it to 2048. The corresponding unit tests
are updated to use this constant as well.

Change-Id: I9aeee6213769a552666af1a4c50b33641f4eeca9
Signed-off-by: Greg Zaverucha <gregz@microsoft.com>