ASACORE-1582 BigNum code returns incorrect types
operator- and div in BigNum return an integer where they should
return BigNum objects. The compiler implicitly calls a constructor
to convert this to a BigNum object since the return type requires it,
but relying on implicit behavior of the compiler is fragile. Instead,
this change refers to existing BigNum objects or calls the constructor
explicitly to create new ones.

BigNum::mod_inv also exhibits this behavior, but is dead code, and so
this change removes it.

Change-Id: I30cee7c24bd30b02192e045ffe0ea6d22844253a
Signed-off-by: Kevin Kane <kkane@microsoft.com>