ASACORE-1593 Change default SRP parameters
The SRP server role gets to choose which parameter set to use, from
the 1024-bit and 1536-bit set. This changes it from the 1024-bit set
to the 1536-bit set, to align with Microsoft and NIST guidance on
keylength. This change also adds a test to make sure the protocol
fails when the passwords do not match.

Change-Id: I4a30cee66d9a1366c09ce14a7dae8773d081ddd3
Signed-off-by: Greg Zaverucha <gregz@microsoft.com>