ASACORE-1419 Updated the Events and Actions sample
This updates the Events and Actions sample so it does not use the
About Service instead it uses the About Feature.

This is needed so we can mark the About Service depricated for Jira
ticket ASACORE-1419 which is about marking the About Service as
depricated. Since we treat warnings as errors the events and actions
sample would not compile once the About Service code is marked as
depricated. So this is required to complete ASACORE-1419.

Change-Id: I8080db4ba300352a32b45df1d6c5ce230217787a
Signed-off-by: George Nash <georgen@qce.qualcomm.com>