ASACORE-1603 Fixed segfault in AboutData::CreatFromXml
When processing xml if a required field was missing it would
attempet to read the value anyway resulting in a segfault inside
the code.

As part of the fixes for this error I also made the CreateFromXml
function handle missing tags more robust.  If an error is encountered
it will try and process the rest of the xml.

Change-Id: Ibdfc4556cf23b7ca5d31fb5fbceeebbe68378ed2
Signed-off-by: George Nash <georgen@qce.qualcomm.com>