ASACORE-1383 avoid confusion in upcoming merge to master
One of the very first patches on the observer feature branch introduced
non-blocking versions of a number of BusAttachment methods
(WhoImplements, AddMatch, etc). The commit was later cherry-picked to
another feature branch, where some additional fixes were applied.

To avoid confusion when the two feature branches are merged to master,
this patch lines up the implementation of the nonblocking methods
between both branches.

Change-Id: Iefeb5cd7a543b9c2efdd0377eb797756cd651a54
Signed-off-by: Meggan Ehret <meggan@qeo.org>