ASACORE-1603 Updated Java CreateFromXml behavior to match C++
While cleaning up NullPointerException found using static analysis
tool I discovered the createFromXml method did not behave the same
in Java as it does in the C++ code.

I adjusted the behavior of the createFromXml so it will try and
process all of the xml tags. Similar to the C++ code.

The createFromXml method now throws its exceptions rather than
changing the exceptions to a BusException. This give the user
more information about the nature of the failure.

In addition I cleaning up the NullPointerException catch block
that is considered bad programming practice.

Change-Id: If06a83cc64326ca4ed3da2db8823b5c97d431b5d
Signed-off-by: George Nash <georgen@qce.qualcomm.com>