ASACORE-1687 AJTest PropChangedTest.PropertyCache_async is failing
The /vmm /vmg compiler flags removal from Windows builds caused
this test to fail or crash on some machines. On my machine, this
test just fails (no crash), so I was able to somewhat repro.

My finding is that the instance pointer generated by the compiler
was not correct as the definitions of the member pointer in AJCore
also needs to see the pragma. Eseentially my "local" fix for /vmm
and /vmg removal was not enough -- I needed to do this at a global
scope.

I verified the affected tests after my new change.

Change-Id: I5f91425283fac6dbf978456102606b3daf32be23
Signed-off-by: Way Vadhanasin <wayakorn.vadhanasin@microsoft.com>