ASACORE-1641 Remove PIN_KEYX authentication mech
The ALLJOYN_PIN_KEYX mechanism is designed for thin clients prior
to the 14.06 release, and is still present in the thick client
implementation.

For most PIN lengths, the PIN can be recovered by observing a
protocol transcript. From the PIN and public nonces, a master secret
is derived, then used to exchange information. A brute force
guessing attack can be mounted on the client_finish message, since
the PIN is the only unknown input. Therefore, a passive adversary
can record the authentication flows then all subsequent sessions,
perform the brute force attack at his leisure, then decrypt all
data. To provide an acceptable level of security, PINs must be
pre-shared keys, having 128-bits of entropy or more. For (base-10)
numeric PINs the length must be at least 39 characters. For
alphanumeric PINs from a 64-character alphabet, the length must be
at least 22 characters.

There is an agreement between MS and QC to remove this
mechanism in favor of SRP when a password must be used.

I removed the AuthMechPIN.cc and its header file, removed
text references to this mechanism, and replaced it with
ALLJOYN_SRP_KEYX wherever possible.

Change-Id: Ie558425920323c697e1e7c26c8857c28639225c6
Signed-off-by: Tolga Acar <tolga@microsoft.com>