ASACORE-1659 Ensure hardcoded GUIDs are not used by remote peers
Maintain a list of hardcoded GUIDs used for storage of local keys,
and reject any attempt by a peer to claim this as their identifying
GUID as a potential attempt to learn something about one of the local
keys. During peer authentication, reject any attempt that uses one of
these GUIDs. If the unlikely event of one of them being randomly
generated happens, spin again.

Added tests to exercise either a malicious client or server attempting
to use one of these GUIDs.

While "in the neighborhood," correct the spelling of ExcerciseOn to
ExerciseOn in AuthListenerECDHETest.

Change-Id: I28b1288100396c05e715a5fffc2317ff7e39e362
Signed-off-by: Kevin Kane <kkane@microsoft.com>