ASACORE-1665 Remove AES-CCM PRF
The pseudorandom function implemented in
Crypto_PseudoRandomFuncitonCCM is non-standard, and may not provide
sufficient cryptographic security (it is used for key derivation,
and key confirmation). Greg's and Tolga's crypto reviews didn't
uncover a weakness, but a more comprehensive review might be required
to evaluate its security by a wider crypto expert audience (Microsoft
Crypto Board). Changing it to a standard function is
preferred, such as a HMAC-based PRF as in NIST SP800-108.

After a discussion with Greg Burns and Phil Nguyen, this PRF can be
removed only if the PIN_KEYX mechanism is removed from the standard
client (ASACORE-1641).

Change-Id: I94785e8dab6dfa106175a5ef674e85126964b1bb
Signed-off-by: Tolga Acar <tolga@microsoft.com>