ASACORE-1713 Remove DBUS_COOKIE_SHA1 authentication mechanism
(DBusCookieSHA1.cc) This protocol is vulnerable to a
man-in-the-middle attack. The protocol works as follows: the
server sends a challenge nonce, the client chooses another nonce
and computes

response = SHA1(srvNonce||":"||clientNonce||":"||cookie)

where cookie is a pre-shared secret. A MITM can relay the server
nonce to the client and then forward the client nonce and
response back to the server.  So on its own this protocol is
weak, but it might provide some security depending on the
context.

There is agreement between Qualcomm and Microsoft as of Feb
27/2015 this auth mechanism is not used by AJ. Scheduled for
removal.

Change-Id: Ia930d6749ef4d54ddb9abe73341bf177cbe7bd53
Signed-off-by: Tolga Acar <tolga@microsoft.com>