AJCORE-1652 Added private copy/assign functions
When a class or struct has dynamic memory it should also
defince a copy constructor and assignment operator to prevent
double freeing of memory.

If the class or struct should never be copied a private version
of these functions must be added to prevent the compiler from auto
generating the functions.

Change-Id: I502dd87e06fd9dd6a4db88a1725c2f3024e30617
Signed-off-by: George Nash <georgen@qce.qualcomm.com>