ASACORE-1716 Remove RSA auth mechanism (part 1)
After discussions on the core mailing list it was decided to remove
the RSA authentication mechanism. This change remove the mechanism
and related crypto code, and updates most of the samples and unit
tests.  The java bindings are not updated in this change.  A second
commit (part 2) will update the alljoyn_java code to work with this
change.

Change-Id: I889391341efcc8be4043a34d65cd95948c468dc1
Signed-off-by: Greg Zaverucha <gregz@microsoft.com>